feat: make Turing a weak dependency (#36, phase 2)
Turing was the most volatile dependency in the project (five permitted minors)
and the largest registry-compat blast radius: a breaking Turing release can
block `Pkg.add("NoLimits")` for a published version while `main` stays green
(#133). As a weak dependency that churn only reaches users who opt in. It also
takes 21 packages and ~0.65s off a default load (4.9s -> 4.26s warm, 274 -> 253
packages).

`MCMC`, `VI`, `MCMCResult` and `VIResult` could not move: `src/` refers to them
in roughly thirty places across uq, serialization, cv, plotting and summaries.
They now live in `estimation/mcmc_types.jl` together with the chain accessors,
`build_fit_result` for chains, and `sample_posterior`; only the code that builds
DynamicPPL models and runs samplers moved to `NoLimitsTuringExt`. `MCMC`'s
`sampler`/`adtype` and `uq`'s `mcmc_sampler`/`mcmc_adtype` default to `nothing`
and resolve to `NUTS(0.75)`/`AutoForwardDiff()` inside the extension.

DynamicPPL is dropped from `[deps]` - inside the extension it is reachable as
`Turing.DynamicPPL` (a weakdep that is not itself a trigger is invisible to an
extension, so the indirection is required). Bijectors and MCMCChains stay direct
deps: `remeasure.jl` and `dev_api.jl` use Bijectors outside the samplers, and
`MCMCChains.Chains` is the documented return type of `get_chain`.

`uq/mcmc_refit.jl` turned out to need nothing from Turing beyond a `using` line
- it reaches Turing only through `fit_model(dm, MCMC(...))`, which is gated - so
it stays in `src/`. `kernel.jl`'s `nothing`-sampler fallback becomes the native
`AdaptiveNoLimitsMH()` rather than `Turing.MH()`, keeping that path working
without Turing.

Gated paths error through `_require_ext` with the actionable message: MCMC, VI,
Turing E-step samplers in SAEM/MCEM, and unlinking variational draws (a
deserialized VI result carries no model and still needs no Turing).
