+++ Project.toml
@@ -13,7 +13,6 @@ DiffEqBase = "2b5f629d-d688-5b77-993f-72d75c75574e"
 DiffEqCallbacks = "459566f4-90b8-5000-8ac3-15dfb0a30def"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
-DynamicPPL = "366bfd00-2699-11ea-058f-f148b4cae6d8"
 ExponentialAction = "e24c0720-ea99-47e8-929e-571b494574d3"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 FunctionChains = "8e6b2b91-af83-483e-ba35-d00930e4cf9b"
@@ -38,7 +37,6 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
-Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
 
 [weakdeps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
@@ -51,6 +49,7 @@ Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 OptimizationNLopt = "4e6fcdb7-1186-4e1f-a706-475e75c168bb"
 SciMLSensitivity = "1ed8b502-d754-442c-8d5d-10ac956f44a1"
 SimpleChains = "de6bee2f-e2f4-4ec7-b6ed-219cc6f6e9e5"
+Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
 
 [extensions]
 NoLimitsCSVExt = "CSV"
@@ -62,6 +61,7 @@ NoLimitsLuxExt = "Lux"
 NoLimitsMakieExt = "Makie"
 NoLimitsSciMLSensitivityExt = "SciMLSensitivity"
 NoLimitsSimpleChainsExt = "SimpleChains"
+NoLimitsTuringExt = "Turing"
 
 [compat]
 Bijectors = "0.15.15, 0.16"
@@ -74,7 +74,6 @@ DiffEqBase = "6.199.0, 7.0.0 - 7.5.5, 7.6"
 DiffEqCallbacks = "4.12.0"
 Distributions = "0.25.123"
 Downloads = "1.6.0"
-DynamicPPL = "0.39.12, 0.41, 0.42"
 ExponentialAction = "0.2.10"
 ForwardDiff = "1"
 FunctionChains = "0.2.3"
