+++ src/NoLimits.jl
@@ -42,8 +42,7 @@ include("estimation/re_batch.jl")
 include("estimation/population_moments.jl")
 include("estimation/mle.jl")
 include("estimation/map.jl")
-include("estimation/mcmc_turing.jl")
-include("estimation/vi_turing.jl")
+include("estimation/mcmc_types.jl")
 include("estimation/laplace.jl")
 include("estimation/focei.jl")
 include("estimation/pooled.jl")
