+++ test/Manifest.toml
@@ -1277,9 +1277,9 @@ weakdeps = ["Adapt", "Random"]
 
 [[deps.FlexiChains]]
 deps = ["AbstractMCMC", "AbstractPPL", "DelimitedFiles", "DimensionalData", "DocStringExtensions", "LinearAlgebra", "MCMCDiagnosticTools", "OrderedCollections", "PrecompileTools", "Printf", "Random", "Statistics", "StatsBase", "Tables"]
-git-tree-sha1 = "d50dcb1b13d446b2738ca5e54a8327963c91b1b4"
+git-tree-sha1 = "1b35c5d07782e74397e0d0fc0c7a2446d207c024"
 uuid = "4a37a8b9-6e57-4b92-8664-298d46e639f7"
-version = "0.6.28"
+version = "0.6.37"
 
     [deps.FlexiChains.extensions]
     FlexiChainsAdvancedHMCExt = ["AdvancedHMC", "DimensionalData"]
@@ -1288,6 +1288,7 @@ version = "0.6.28"
     FlexiChainsInferenceObjectsExt = ["InferenceObjects", "DimensionalData", "OrderedCollections"]
     FlexiChainsMCMCChainsExt = ["MCMCChains", "OrderedCollections"]
     FlexiChainsMakieExt = ["Makie", "StatsBase", "KernelDensity"]
+    FlexiChainsMonteCarloMeasurementsExt = ["MonteCarloMeasurements", "OrderedCollections"]
     FlexiChainsPairPlotsExt = ["PairPlots", "Makie"]
     FlexiChainsPigeonsDynamicPPLExt = ["Pigeons", "DynamicPPL", "AbstractMCMC"]
     FlexiChainsPigeonsExt = ["Pigeons"]
@@ -1308,6 +1309,7 @@ version = "0.6.28"
     KernelDensity = "5ab0869b-81aa-558d-bb23-cbf5423bbe9b"
     MCMCChains = "c7f686f2-ff18-58e9-bc7b-31028e88f75d"
     Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
+    MonteCarloMeasurements = "0987c9cc-fe09-11e8-30f0-b96dd679fdca"
     PairPlots = "43a3c2be-4208-490b-832a-a21dcd55d7da"
     Pigeons = "0eb8d820-af6a-4919-95ae-11206f830c31"
     PosteriorDB = "1c4bc282-d2f5-44f9-b6d1-8c4424a23ad4"
@@ -2516,11 +2518,11 @@ uuid = "ca575930-c2e3-43a9-ace4-1e988b2c1908"
 version = "1.3.0"
 
 [[deps.NoLimits]]
-deps = ["Bijectors", "ComponentArrays", "DataFrames", "DataInterpolations", "DiffEqBase", "DiffEqCallbacks", "Distributions", "Downloads", "DynamicPPL", "ExponentialAction", "ForwardDiff", "FunctionChains", "Functors", "KernelDensity", "LineSearches", "LinearAlgebra", "Logging", "MCMCChains", "Optimisers", "Optimization", "OptimizationOptimJL", "OrdinaryDiffEq", "ProgressMeter", "Random", "Roots", "RuntimeGeneratedFunctions", "SciMLBase", "SciMLStructures", "SpecialFunctions", "StaticArrays", "Statistics", "StatsFuns", "Symbolics", "Turing"]
+deps = ["Bijectors", "ComponentArrays", "DataFrames", "DataInterpolations", "DiffEqBase", "DiffEqCallbacks", "Distributions", "Downloads", "ExponentialAction", "ForwardDiff", "FunctionChains", "Functors", "KernelDensity", "LineSearches", "LinearAlgebra", "Logging", "MCMCChains", "Optimisers", "Optimization", "OptimizationOptimJL", "OrdinaryDiffEq", "ProgressMeter", "Random", "Roots", "RuntimeGeneratedFunctions", "SciMLBase", "SciMLStructures", "SpecialFunctions", "StaticArrays", "Statistics", "StatsFuns", "Symbolics"]
 path = ".."
 uuid = "1d491b74-35a5-408f-ae19-cf8524c3769d"
 version = "0.2.2"
-weakdeps = ["CSV", "Copulas", "JLD2", "Latexify", "LikelihoodProfiler", "Lux", "Makie", "OptimizationNLopt", "SciMLSensitivity", "SimpleChains"]
+weakdeps = ["CSV", "Copulas", "JLD2", "Latexify", "LikelihoodProfiler", "Lux", "Makie", "OptimizationNLopt", "SciMLSensitivity", "SimpleChains", "Turing"]
 
     [deps.NoLimits.extensions]
     NoLimitsCSVExt = "CSV"
@@ -2532,6 +2534,7 @@ weakdeps = ["CSV", "Copulas", "JLD2", "Latexify", "LikelihoodProfiler", "Lux", "
     NoLimitsMakieExt = "Makie"
     NoLimitsSciMLSensitivityExt = "SciMLSensitivity"
     NoLimitsSimpleChainsExt = "SimpleChains"
+    NoLimitsTuringExt = "Turing"
 
 [[deps.NonlinearSolve]]
 deps = ["ADTypes", "ArrayInterface", "BracketingNonlinearSolve", "CommonSolve", "ConcreteStructs", "DifferentiationInterface", "FastClosures", "FiniteDiff", "ForwardDiff", "LineSearch", "LinearAlgebra", "LinearSolve", "NonlinearSolveBase", "NonlinearSolveFirstOrder", "NonlinearSolveQuasiNewton", "NonlinearSolveSpectralMethods", "PrecompileTools", "Preferences", "Reexport", "SciMLBase", "Setfield", "SimpleNonlinearSolve", "StaticArraysCore", "SymbolicIndexingInterface"]
